# (#3039) - WIP DON'T MERGE - use a bloomfilter for idb