--- a/lib/adapters/idb/idb-bulk-docs.js
+++ b/lib/adapters/idb/idb-bulk-docs.js
@@ -118,7 +118,13 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
         checkDone(); // skip local docs
         continue;
       }
-      var req = docStore.get(docInfo.metadata.id);
+      var id = docInfo.metadata.id;
+      if (!api._docIdsFilter.test(id)) { // doesn't exist
+        api._docIdsFilter.add(id); // for races with simultaneous bulkDocs
+        checkDone();
+        continue;
+      }
+      var req = docStore.get(id);
       req.onsuccess = readMetadata;
     }
   }
@@ -128,8 +134,8 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
       return;
     }
 
-    Changes.notify(api._name);
     api._docCount = -1; // invalidate
+    Changes.notify(api._name);
     callback(null, results);
   }