--- a/lib/adapters/idb/idb-utils.js
+++ b/lib/adapters/idb/idb-utils.js
@@ -241,3 +241,16 @@ exports.openTransactionSafely = function (idb, stores, mode) {
     };
   }
 };
+
+// simplified cursor API
+exports.openCursor = function (store, range, onGetCursor, onDone) {
+  var cursor = range ? store.openCursor(range) : store.openCursor();
+  cursor.onsuccess = function (e) {
+    cursor = e.target.result;
+    if (!cursor) {
+      return onDone();
+    }
+    onGetCursor(cursor);
+    cursor.continue();
+  };
+};