--- a/lib/adapters/idb/idb.js
+++ b/lib/adapters/idb/idb.js
@@ -1,5 +1,6 @@
 'use strict';
 
+var BloomFilter = require('bloomfilter').BloomFilter;
 var utils = require('../../utils');
 var merge = require('../../merge');
 var errors = require('../../deps/errors');
@@ -28,6 +29,7 @@ var postProcessAttachments = idbUtils.postProcessAttachments;
 var readBlobData = idbUtils.readBlobData;
 var taskQueue = idbUtils.taskQueue;
 var openTransactionSafely = idbUtils.openTransactionSafely;
+var openCursor = idbUtils.openCursor;
 
 var cachedDBs = {};
 var blobSupportPromise;
@@ -54,6 +56,7 @@ function init(api, opts, callback) {
   api._docCount = -1;
   api._blobSupport = null;
   api._name = name;
+  api._docIdsFilter = null;
 
   // called when creating a fresh new database
   function createSchema(db) {
@@ -294,6 +297,12 @@ function init(api, opts, callback) {
     var err;
     var txn;
     opts = utils.clone(opts);
+
+    // check the cache first for perf
+    if (!api._docIdsFilter.test(id)) {
+      return callback(errors.error(errors.MISSING_DOC, 'missing'));
+    }
+
     if (opts.ctx) {
       txn = opts.ctx;
     } else {
@@ -827,6 +836,12 @@ function init(api, opts, callback) {
 
 
   api._getLocal = function (id, callback) {
+
+    // check the cache first for perf
+    if (!api._docIdsFilter.test(id)) {
+      return callback(errors.error(errors.MISSING_DOC));
+    }
+
     var txnResult = openTransactionSafely(idb, [LOCAL_STORE], 'readonly');
     if (txnResult.error) {
       return callback(txnResult.error);
@@ -858,7 +873,12 @@ function init(api, opts, callback) {
       doc._rev = '0-1';
     } else {
       doc._rev = '0-' + (parseInt(oldRev.split('-')[1], 10) + 1);
+      if (!api._docIdsFilter.test(doc._id)) {
+        // doc doesn't exist
+        return callback(errors.error(errors.REV_CONFLICT));
+      }
     }
+    api._docIdsFilter.add(doc._id); // prevent from using cache later
 
     var tx = opts.ctx;
     var ret;
@@ -945,6 +965,7 @@ function init(api, opts, callback) {
     idb = cached.idb;
     instanceId = cached.instanceId;
     api._blobSupport = cached.blobSupport;
+    api._docIdsFilter = cached.docIdsFilter;
     process.nextTick(function () {
       callback(null, api);
     });
@@ -1009,50 +1030,83 @@ function init(api, opts, callback) {
       delete cachedDBs[name];
     };
 
-    var txn = idb.transaction([META_STORE, DETECT_BLOB_SUPPORT_STORE],
+    var txn = idb.transaction([
+        META_STORE,
+        DETECT_BLOB_SUPPORT_STORE,
+        DOC_STORE,
+        LOCAL_STORE
+      ],
       'readwrite');
 
     var req = txn.objectStore(META_STORE).get(META_STORE);
 
     req.onsuccess = function (e) {
 
-      var checkSetupComplete = function () {
-        if (api._blobSupport === null || !idStored) {
+      function checkSetupComplete() {
+        if (api._blobSupport === null || !idStored || !api._docIdsFilter) {
           return;
         } else {
           cachedDBs[name] = {
             idb: idb,
             instanceId: instanceId,
             blobSupport: api._blobSupport,
+            docIdsFilter:  api._docIdsFilter,
             loaded: true
           };
           callback(null, api);
         }
-      };
+      }
 
-      var meta = e.target.result || {id: META_STORE};
-      if (name  + '_id' in meta) {
-        instanceId = meta[name + '_id'];
-        idStored = true;
-        checkSetupComplete();
-      } else {
-        instanceId = utils.uuid();
-        meta[name + '_id'] = instanceId;
-        txn.objectStore(META_STORE).put(meta).onsuccess = function () {
+      function storeId() {
+        var meta = e.target.result || {id: META_STORE};
+        if (name  + '_id' in meta) {
+          instanceId = meta[name + '_id'];
           idStored = true;
           checkSetupComplete();
-        };
+        } else {
+          instanceId = utils.uuid();
+          meta[name + '_id'] = instanceId;
+          txn.objectStore(META_STORE).put(meta).onsuccess = function () {
+            idStored = true;
+            checkSetupComplete();
+          };
+        }
       }
 
-      if (!blobSupportPromise) {
-        // make sure blob support is only checked once
-        blobSupportPromise = checkBlobSupport(txn, idb);
+      function doCheckBlobSupport() {
+        if (!blobSupportPromise) {
+          // make sure blob support is only checked once
+          blobSupportPromise = checkBlobSupport(txn, idb);
+        }
+
+        blobSupportPromise.then(function (val) {
+          api._blobSupport = val;
+          checkSetupComplete();
+        });
       }
 
-      blobSupportPromise.then(function (val) {
-        api._blobSupport = val;
-        checkSetupComplete();
-      });
+      function fetchDocIds() {
+        var docIds = new BloomFilter(8192, 16);
+        var numDone = 0;
+
+        function onGetCursor(cursor) {
+          docIds.add(cursor.primaryKey);
+        }
+
+        function checkDone() {
+          if (++numDone === 2) {
+            api._docIdsFilter = docIds;
+            checkSetupComplete();
+          }
+        }
+
+        openCursor(txn.objectStore(DOC_STORE), null, onGetCursor, checkDone);
+        openCursor(txn.objectStore(LOCAL_STORE), null, onGetCursor, checkDone);
+      }
+
+      storeId();
+      doCheckBlobSupport();
+      fetchDocIds();
     };
   };