When we're doing `bulkDocs`, one of the things that slows us down a lot is checking for existing documents by id. This is especially unfortunate for generation-1 docs, because the docs don't exist and so the `get()` is pointless.

This adds a bloom filter to keep track of doc ids, so that we can bail out early if we know a document has never been inserted before. 

Performance impact (temp-views test, 5 iterations):

    Firefox 37: 52444ms -> 49611ms (5% improvement)
    Chrome 40: 65949ms -> 46770ms (29% improvement)
    IE 11: 57911ms -> 54317ms (6% improvement)

Footprint size change:  46170 before, 47136 after (min+gz)

I can eventually implement this for WebSQL and LevelDB as well. Opening this PR early for feedback.