@nolanlawson 2015-02-01
small fixup