@nolanlawson 2015-02-01
Well that's weird. The new test is failing in CouchDB master because apparently you can use two simultaneous `bulkDocs` to PUT the same `_local` doc, and you'll get a 201 for each one. I understand the whole clustering thing, but shouldn't that not apply to `_local` docs @willholley?

If nothing else, even the original test is flakey in CouchDB master, so I can just add an `isCouchMaster()` check to both.