@nolanlawson 2015-02-01
added the `isCouchMaster` test.