@nolanlawson 2015-02-02
Hm, not improving the perf for WebSQL, which makes me nervous. Also we might want to play around with different cache sizes: 8192 actually gives us a really high false positive rate.