@willholley 2015-02-02
@nolanlawson - _local docs do not have MVCC and are not replicated around a
cluster. If you write two copies simultaneously then I think it will be
last write wins (which might not be the same as last request wins). This
actually might cause some strange behaviour in a cluster  (a discussion for
#couchdb-dev) but my understanding is that's how it's worked in Cloudant
forever...

On 1 February 2015 at 23:20, Nolan Lawson <notifications@github.com> wrote:

> Well that's weird. The new test is failing in CouchDB master because
> apparently you can use two simultaneous bulkDocs to PUT the same _local
> doc, and you'll get a 201 for each one. I understand the whole clustering
> thing, but shouldn't that not apply to _local docs @willholley
> <https://github.com/willholley>?
>
> If nothing else, even the original test is flakey in CouchDB master, so I
> can just add an isCouchMaster() check to both.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/pouchdb/pouchdb/pull/3485#issuecomment-72390790>.
>