@nolanlawson 2015-02-07
Closing because I don't want the extra complexity in exchange for modest perf gains. LevelDB already has a bloom filter, let the browser vendors make IndexedDB faster.